==> src/analyses/01_Make_DEG_Lists.Rmd <==
---
title: "Get DEGs list with plots"
author: "Chris Mancuso"
date: "2023-09-27"
output: 
  html_document:
    code_folding: "hide"
    toc: true
    toc_depth: 3
    toc_float: true
theme: "cerulean"
---

# Lood in all the needed pacakges

```{r, message=FALSE, warning=FALSE}
library(DESeq2)
library(httr)
library(tximport)
library(plotly)
library(readxl)
library(ashr)
library(repr)
library(pheatmap)
library(RColorBrewer)
library(DT)
library(patchwork)
library(EnhancedVolcano)
library(clusterProfiler)
library(org.Ss.eg.db)
library(org.Hs.eg.db)
library(ReactomePA)
library(UpSetR)
library(tidyverse)
```

```{r, message=FALSE, warning=FALSE}
myLFC <- 0.59
mypadj <- 0.05
fp_save <- "../../results/Make_DEG_Lists/"
```


```{r, message=FALSE, warning=FALSE}

make_ortho_df <- function(){
# Load the pig to human gene converting file from BioMart
fpmart <- "../../data/orthologs/"
dfmart <- read_tsv(paste0(fpmart,"pig_mart_export.txt")) %>% rename(`GeneID` = 1) %>%
filter(`Human orthology confidence [0 low, 1 high]` > 0.5) %>%
dplyr::select(`GeneID`,`Gene name`,`Human gene stable ID`)
# Load in file with more human gene info
dfmart_hs <- read_tsv(paste0(fpmart,"human_mart_export.txt")) %>%
    dplyr::select(-`Gene Synonym`) %>%
  distinct() %>%
  rename(`Human gene stable ID` = 1,`Human gene name` = 2, 
         `Human gene description` = 3)
dfgene <- left_join(dfmart,dfmart_hs, by="Human gene stable ID")
return(dfgene)
}
dfgene <- make_ortho_df()

make_deseq2_obj <- function(deseq_type){
  fp_count <- "../../data/mapped_reads"
  fp_count1b <- "../../data/mapped_reads/run1"
  fp_count2b <- "../../data/mapped_reads/run2"
  fp_meta <- "../../data/"
  dirs1 <-  list.dirs(fp_count1b,full.names = F)
  dirs1 <- dirs1[-1]
  files1 <- file.path(paste0(fp_count1b,"/",dirs1),"quant.sf")
  dirs2 <-  list.dirs(fp_count2b,full.names = F)
  dirs2 <- dirs2[-1]
  files2 <- file.path(paste0(fp_count2b,"/",dirs2),"quant.sf")
  files <- c(files1,files2)
  # read in txt2gene file
  txt2gene <- read_tsv(paste0(fp_count,"/salmon_tx2gene.tsv"),col_names = F)
  # read in metadata
  coldata <- read_csv(paste0(fp_meta,"Sample info + gender with heart.csv")) %>%
    filter_all(any_vars(!is.na(.)))
  coldata <- column_to_rownames(coldata,var="SampleName")
  # order the metadata the same way the counts were read in
  dirs2b <- paste0("X",dirs2)
  coldata <- coldata[ c(dirs1,dirs2b), ]
  if (deseq_type == "AllTogether"){
    # don't do anything
  } else {
    inds_tmp <- which(coldata$Tissue==deseq_type)
    files <- files[inds_tmp]
    coldata <- coldata[inds_tmp,]
  }
  txi <- tximport(files, type="salmon", tx2gene=txt2gene)
  ddsgrp <- DESeqDataSetFromTximport(txi,
                                     colData = coldata,
                                     design = ~ Group)
  ddsgrp <- DESeq(ddsgrp)
  return(list(ddsgrp,coldata))
}

make_results <- function(deseq_obj,group1,group2,atissue){
  res <- results(deseq_obj,alpha=myLFC,contrast=c("Group",
                                                  sprintf("%s%s",group1,atissue),
                                                  sprintf("%s%s",group2,atissue)))
  resLFC <- lfcShrink(deseq_obj,
                      res = res,
                      type="ashr")
  # get gene names from count file generated by nf-core pipeline
  fp_count <- "../../data/mapped_reads"
  symdf <- read_tsv(paste0(fp_count,"/salmon.merged.gene_counts_length_scaled.tsv"))
  resLFC$GeneName <-  symdf$gene_name
  resLFC$GeneID <-  symdf$gene_id
  resOrderLFC <- resLFC[order(resLFC$padj),]
  return(resOrderLFC)
}

make_heatmap <- function(deseq_obj,group1,group2,coldata, gene_table1,gene_table2,tiss,scale_type){
  # load a count table to get the gene IDs
  fp_count <- "../../data/mapped_reads"
  symdf <- read_tsv(paste0(fp_count,"/salmon.merged.gene_counts_length_scaled.tsv"))
  # create counts assay
  myassay <- counts(deseq_obj,normalized=TRUE)
  row.names(myassay) <- symdf$gene_id
  # find the up and down regulated genes
  upgenes <- gene_table1 %>% pull(`GeneID`) %>% na.omit() %>% unique()
  downgenes <- gene_table2 %>% pull(`GeneID`) %>% na.omit() %>% unique()
  allgenes <- union(upgenes,downgenes)
  all_genes_inds <- which(symdf$gene_id %in% allgenes)
  # subset the assay to those genes
  myassay <- myassay[all_genes_inds,]
  # get the correct columns 
  grp_names <- coldata$Group
  con1_inds <- which(coldata$Group == paste0(group1,tiss))
  con2_inds <- which(coldata$Group == paste0(group2,tiss))
  all_con_inds <- c(con1_inds,con2_inds)
  # subset the assay for correct labels
  myassay <- myassay[,all_con_inds]
  # scale the data
  if (scale_type == "Relative"){
    myassay <- t(scale(t(myassay))) %>% na.omit()
  } else if (scale_type == "CPM"){
    myassay <- log2(myassay + 1) %>% na.omit()
  }
  # make label df
  df_label <- as.data.frame(colData(deseq_obj))
  df_label  <- df_label %>% filter(Group %in% c(paste0(group1,tiss),paste0(group2,tiss)))
  df_label <- df_label[,"Group", drop=FALSE]
  # all_genes_inds <- which(symdf$gene_name %in% allgenes)
  if (dim(myassay)[1]>2){
    myrowclus <- TRUE
  } else {
    myrowclus <- FALSE
  }
  hm <- pheatmap(myassay, cluster_rows=myrowclus, show_rownames=FALSE,
                          cluster_cols=FALSE, annotation_col=df_label)
  return(hm)
}

make_GO_tables <- function(gene_table,resOrderLFC){ 
  sig_genes <- gene_table %>% drop_na(`Human gene stable ID`) %>% pull(var=`Human gene stable ID`) %>% na.omit() %>% unique()
  uni_genes <- left_join(resOrderLFC,dfgene, by="GeneID") %>% drop_na(`Human gene stable ID`) %>% pull(var=`Human gene stable ID`) %>% na.omit()
  ego <- enrichGO(gene          = sig_genes,
                  universe      = uni_genes,
                  OrgDb         = org.Hs.eg.db,
                  keyType       = 'ENSEMBL',
                  ont           = "BP",
                  minGSSize     = 5,
                  maxGSSize     = 500,
                  pAdjustMethod = "BH",
                  pvalueCutoff  = 0.05,
                  qvalueCutoff  = 0.2,
                  readable      = TRUE)
  if (!is.null(ego)){
  ego <- filter(ego, p.adjust < 0.05)
  }
  return(ego)
}

make_reactome_tables <- function(gene_table,resOrderLFC){
  sig_genes <- gene_table %>% drop_na(`Human gene stable ID`) %>% pull(var=`Human gene stable ID`) %>% na.omit() %>% unique()
  if (length(sig_genes)>0){
    geneids <- stack(mget(sig_genes, org.Hs.egENSEMBL2EG, ifnotfound = NA)) %>% na.omit() %>% pull(values)
  } else {
    geneids <- c()
  }
  uni_genes <- left_join(resOrderLFC,dfgene, by="GeneID") %>% drop_na(`Human gene stable ID`) %>% pull(var=`Human gene stable ID`) %>% na.omit()
  uniids <- stack(mget(uni_genes, org.Hs.egENSEMBL2EG, ifnotfound = NA)) %>% na.omit() %>% pull(values)
  eRP <- enrichPathway(gene          = geneids, 
                       universe      = uniids,
                       organism     = "human",
                       minGSSize     = 5,
                       maxGSSize     = 500,
                       pAdjustMethod = "BH",
                       pvalueCutoff  = 0.05,
                       qvalueCutoff  = 0.2)
  if (!is.null(eRP)){
    eRP <- filter(eRP, p.adjust < 0.05)
    eRP <- setReadable(eRP, OrgDb = org.Hs.eg.db, keyType="ENTREZID")
  }
  return(eRP)
}

make_kegg_tables <- function(gene_table,resOrderLFC){
  sig_genes <- gene_table %>% drop_na(`Human gene stable ID`) %>% pull(var=`Human gene stable ID`) %>% na.omit() %>% unique()
  if (length(sig_genes)>0){
    geneids <- stack(mget(sig_genes, org.Hs.egENSEMBL2EG, ifnotfound = NA)) %>% na.omit() %>% pull(values)
  } else {
    geneids <- c()
  }
  uni_genes <- left_join(resOrderLFC,dfgene, by="GeneID") %>% drop_na(`Human gene stable ID`) %>% pull(var=`Human gene stable ID`) %>% na.omit()
  uniids <- stack(mget(uni_genes, org.Hs.egENSEMBL2EG, ifnotfound = NA)) %>% na.omit() %>% pull(values)
  kegg <- enrichKEGG(gene          = geneids,
                     universe      = uniids,
                     organism      = 'hsa',
                     minGSSize     = 5,
                     maxGSSize     = 500,
                     pAdjustMethod = "BH",
                     pvalueCutoff  = 0.05,
                     qvalueCutoff  = 0.2)
  if (!is.null(kegg)){
    kegg <- filter(kegg, p.adjust < 0.05)
    kegg <- setReadable(kegg, OrgDb = org.Hs.eg.db, keyType="ENTREZID")
  }
  return(kegg)
}

save_upset_plot <- function(x, fn){
  my_plot <- upset(fromList(x), order.by = "freq", nsets= 10)
  pdf(file = fn, height = 7, width = 7)
  print(my_plot)  ## use print here
  dev.off()
}

printsig <- function(x) format(x, digits=2)

get_ortho_info <- function(gene_table){
  mylines <- c()
  num_sig_pig <- length(gene_table %>% pull("GeneID") %>% na.omit() %>% unique())
  mylines <- c(mylines, paste0("The number of significant pig genes is ",num_sig_pig))
  num_sig_pig_with_human <- length(gene_table %>% filter(!is.na(`Human gene stable ID`)) %>% pull("GeneID") %>% na.omit() %>% unique())
  num_sig_pig_no_human <- num_sig_pig - num_sig_pig_with_human
  mylines <- c(mylines, paste0("The number of significant pig genes that don't have any humna ortholog information is ",num_sig_pig_no_human))
  pig_genes_more_one_human <- as.data.frame(table(gene_table$GeneName)) %>% filter(Freq > 1) %>% pull(Var1)
  mylines <- c(mylines, paste0("The number of significant pig genes that map to more than one human gene is ",length(pig_genes_more_one_human)))
  for (apiggene in pig_genes_more_one_human){
    multi_maps <- gene_table %>% filter(GeneName == apiggene) %>% pull(`Human gene name`)
    multi_maps <- paste0(multi_maps, collapse = ', ')
    mylines <- c(mylines, paste0(apiggene, " maps to ",multi_maps))
  }
  mylines <- c(mylines, paste0("\n"))
  return(mylines)
}

get_summary <- function(object_full,up_gene_table,down_gene_table){
  object <- as.data.frame(object_full)
  num_nonzero <- dim(filter(object,object$baseMean>0))[1]
  mylines <-c()
  # set some inputs
  mylines <-c(mylines, "Set some inputs")
  mylines <-c(mylines, paste0("The padj threshold is: ",mypadj))
  mylines <-c(mylines, paste0("The LFC threshold is: ",myLFC))
  if (is.null(metadata(object_full)$filterThreshold)) {
  ft <- 0
  } else {
  ft <- round(metadata(object_full)$filterThreshold)
  }
  mylines <-c(mylines, paste0("The filter threshold is: ",ft,"  \n"))
  # Get information back from DESEQ2 results
  mylines <-c(mylines, "Get information back from DESEQ2 results")
  mylines <-c(mylines, paste0("The number of genes that aren't all zeros is ",num_nonzero))
  num_up <- dim(filter(object,object$log2FoldChange>myLFC & object$padj<mypadj))[1]
  num_down <- dim(filter(object,object$log2FoldChange<(-myLFC) & object$padj<mypadj))[1]
  mylines <-c(mylines, paste0("The number of significant up regulated genes is ",num_up,"  [",printsig(num_up/num_nonzero*100),"%]"))
  mylines <-c(mylines, paste0("The number of significant down regulated genes is ",num_down,"  [",printsig(num_down/num_nonzero*100),"%]"))
  num_outliers <- dim(filter(object,object$baseMean > 0 & is.na(object$pvalue)))[1]
  mylines <-c(mylines, paste0("The number of cookCutoff outliers is ",num_outliers,"  [",printsig(num_outliers/num_nonzero*100),"%]"))
  num_filters <- dim(filter(object,!is.na(object$pvalue) & is.na(object$padj)))[1]
  mylines <-c(mylines, paste0("The number of independentFiltering genes (i.e. low counts) is ",num_filters,"  [",printsig(num_filters/num_nonzero*100),"%]","  \n"))
  # get ortholog info that effects ORA
  mylines <- c(mylines, "Get ortho information for up sig genes")
  if (dim(up_gene_table)[1]>0){
    mylines <- c(mylines, get_ortho_info(up_gene_table))
  }
  mylines <- c(mylines, "Get ortho information for down sig genes")
  if (dim(down_gene_table)[1]>0){
    mylines <- c(mylines, get_ortho_info(down_gene_table))
  }
  
  return(mylines)
}
```

```{r, message=FALSE, warning=FALSE}
mygroups <- list(list("CPBnodrug","Controlnodrug"),list("CPBLind","Controlnodrug"),list("CPBLind","CPBnodrug"))
mytissues <- c("Ileum","Liver","Heart","Kidney","Lung")

for (agroup in mygroups){ 
  mygroup1 <- agroup[[1]]
  mygroup2 <- agroup[[2]]
  upset_genes <- list()
  upset_gos <- list()
  for (atissue in mytissues){
    my_save_string <- paste(mygroup1,mygroup2,atissue,sep="_")
    fun_ret <- make_deseq2_obj(atissue)
    # make the deseq objects
    ddsgrp <- fun_ret[[1]]
    col_data <- fun_ret[[2]]
    resOrderLFC_full <- make_results(ddsgrp,mygroup1,mygroup2,atissue)
    resOrderLFC <- as.data.frame(resOrderLFC_full)
    # make volcano plot
    p_vol <- plot(EnhancedVolcano(resOrderLFC,lab = resOrderLFC$GeneName,x = 'log2FoldChange',y = 'padj',ylab = '-log10(FDR)',pCutoff = mypadj, FCcutoff = myLFC))
    ggsave(paste0(fp_save,my_save_string,"_volcano.pdf"),plot=p_vol,device="pdf",width=7,height=7)
    # do up gene things
    up_gene_table <- resOrderLFC %>% filter(resOrderLFC$log2FoldChange>myLFC & resOrderLFC$padj<mypadj)
    up_gene_table <- left_join(up_gene_table,dfgene)
    if (dim(up_gene_table)[1]>0){
      write_csv(up_gene_table, paste0(fp_save,my_save_string,"_up-regulated.csv"))
      upset_genes[paste0(atissue,"-up")] <- list(up_gene_table %>% pull("GeneID") %>% na.omit() %>% unique())
      ego_up <- make_GO_tables(up_gene_table,resOrderLFC)
      if (dim(as.data.frame(ego_up))[1]>0 & !is.null(ego_up)){
        upset_gos[paste0(atissue,"-up")] <- list(as.data.frame(ego_up) %>% pull("ID") %>% na.omit() %>% unique())
        write_csv(as.data.frame(ego_up), paste0(fp_save,my_save_string,"_up-go.csv"))
        ggsave(paste0(fp_save,my_save_string,"_up-go-dotplot.pdf"),plot=dotplot(ego_up, showCategory=10),device="pdf",width=6,height=6)
        ego_up_simp <- clusterProfiler::simplify(ego_up, cutoff=0.7, by="p.adjust", select_fun=min)
        if (dim(as.data.frame(ego_up_simp))[1]>0 & !is.null(ego_up_simp)){
          write_csv(as.data.frame(ego_up_simp), paste0(fp_save,my_save_string,"_up-go-simp-0.7.csv"))
          ggsave(paste0(fp_save,my_save_string,"_up-go-simp-0.7-dotplot.pdf"),plot=dotplot(ego_up_simp, showCategory=10),device="pdf",width=6,height=6)
        }
        ego_up_simp <- clusterProfiler::simplify(ego_up, cutoff=0.5, by="p.adjust", select_fun=min)
        if (dim(as.data.frame(ego_up_simp))[1]>0 & !is.null(ego_up_simp)){
          write_csv(as.data.frame(ego_up_simp), paste0(fp_save,my_save_string,"_up-go-simp-0.5.csv"))
          ggsave(paste0(fp_save,my_save_string,"_up-go-simp-0.5-dotplot.pdf"),plot=dotplot(ego_up_simp, showCategory=10),device="pdf",width=6,height=6)
        }
      }
      erp_up <- make_reactome_tables(up_gene_table,resOrderLFC)
      if (dim(as.data.frame(erp_up))[1]>0 & !is.null(erp_up)){
        write_csv(as.data.frame(erp_up), paste0(fp_save,my_save_string,"_up-reactome.csv"))
        ggsave(paste0(fp_save,my_save_string,"_up-reactome-dotplot.pdf"),plot=dotplot(erp_up, showCategory=10),device="pdf",width=6,height=6)
      }
      keg_up <- make_kegg_tables(up_gene_table,resOrderLFC)
      if (dim(as.data.frame(keg_up))[1]>0 & !is.null(keg_up)){
        write_csv(as.data.frame(keg_up), paste0(fp_save,my_save_string,"_up-kegg.csv"))
        ggsave(paste0(fp_save,my_save_string,"_up-kegg-dotplot.pdf"),plot=dotplot(keg_up, showCategory=10),device="pdf",width=6,height=6)
      }
    }
    # do down gene things
    down_gene_table <- resOrderLFC %>% filter(resOrderLFC$log2FoldChange<(-myLFC) & resOrderLFC$padj<mypadj)
    down_gene_table <- left_join(down_gene_table,dfgene)
    if (dim(down_gene_table)[1]>0){
      write_csv(down_gene_table,paste0(fp_save,my_save_string,"_down-regulated.csv"))
      upset_genes[paste0(atissue,"-down")] <- list(down_gene_table %>% pull("GeneID") %>% na.omit() %>% unique())
      ego_down <- make_GO_tables(down_gene_table,resOrderLFC)
      if (dim(as.data.frame(ego_down))[1]>0 & !is.null(ego_down)){
        upset_gos[paste0(atissue,"-down")] <- list(as.data.frame(ego_down) %>% pull("ID") %>% na.omit() %>% unique())
        write_csv(as.data.frame(ego_down), paste0(fp_save,my_save_string,"_down-go.csv"))
        ggsave(paste0(fp_save,my_save_string,"_down-go-dotplot.pdf"),plot=dotplot(ego_down, showCategory=10),device="pdf",width=6,height=6)
        ego_down_simp <- clusterProfiler::simplify(ego_down, cutoff=0.7, by="p.adjust", select_fun=min)
        if (dim(as.data.frame(ego_down_simp))[1]>0 & !is.null(ego_down_simp)){
          write_csv(as.data.frame(ego_down_simp), paste0(fp_save,my_save_string,"_down-go-simp-0.7.csv"))
          ggsave(paste0(fp_save,my_save_string,"_down-go-simp-0.7-dotplot.pdf"),plot=dotplot(ego_down_simp, showCategory=10),device="pdf",width=6,height=6)
        }
        ego_down_simp <- clusterProfiler::simplify(ego_down, cutoff=0.5, by="p.adjust", select_fun=min)
        if (dim(as.data.frame(ego_down_simp))[1]>0 & !is.null(ego_down_simp)){
          write_csv(as.data.frame(ego_down_simp), paste0(fp_save,my_save_string,"_down-go-simp-0.5.csv"))
          ggsave(paste0(fp_save,my_save_string,"_down-go-simp-0.5-dotplot.pdf"),plot=dotplot(ego_down_simp, showCategory=10),device="pdf",width=6,height=6)
        }
      }
      erp_down <- make_reactome_tables(down_gene_table,resOrderLFC)
      if (dim(as.data.frame(erp_down))[1]>0 & !is.null(erp_down)){
        write_csv(as.data.frame(erp_down), paste0(fp_save,my_save_string,"_down-reactome.csv"))
        ggsave(paste0(fp_save,my_save_string,"_down-reactome-dotplot.pdf"),plot=dotplot(erp_down, showCategory=10),device="pdf",width=6,height=6)
      }
      keg_down <- make_kegg_tables(down_gene_table,resOrderLFC)
      if (dim(as.data.frame(keg_down))[1]>0 & !is.null(keg_down)){
        write_csv(as.data.frame(keg_down), paste0(fp_save,my_save_string,"_down-kegg.csv"))
        ggsave(paste0(fp_save,my_save_string,"_down-kegg-dotplot.pdf"),plot=dotplot(keg_down, showCategory=10),device="pdf",width=6,height=6)
      }
    }
    if (dim(up_gene_table)[1] + dim(down_gene_table)[1] > 0){
      # make some heatmaps
      p_hm <- make_heatmap(ddsgrp,mygroup1,mygroup2,col_data, up_gene_table, down_gene_table, atissue, "Relative")
      ggsave(paste0(fp_save,my_save_string,"_relativeheatmap.pdf"),plot=p_hm,device="pdf",width=7,height=7)
      p_hm <- make_heatmap(ddsgrp,mygroup1,mygroup2,col_data, up_gene_table, down_gene_table, atissue, "CPM")
      ggsave(paste0(fp_save,my_save_string,"_cpmheatmap.pdf"),plot=p_hm,device="pdf",width=7,height=7)
      # make the summary info
      my_lines <- get_summary(resOrderLFC_full,up_gene_table,down_gene_table)
      write_lines(my_lines, paste0(fp_save,my_save_string,"_summary.txt"))
    }
  }
  save_upset_plot(upset_genes,paste0(fp_save,mygroup1,"_",mygroup2,"_upset-genes.pdf"))
  save_upset_plot(upset_gos,paste0(fp_save,mygroup1,"_",mygroup2,"_upset-gos.pdf"))
}
```
